# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyApp API",
    "version": "1.0.0",
    "description": "API documentation generated from Next.js API routes"
  },
  "paths": {
    "/api/articles": {
      "get": {
        "summary": "Get articles (optionally filtered by sport)",
        "description": "Retrieve all articles from the database.   Optionally filter results by `sport` query parameter (case-insensitive).   If `sport=all` or no parameter is provided, all articles are returned.\n",
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sport",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter articles by sport slug (case-insensitive). Use \"all\" or omit to return everything.",
            "example": "soccer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "MongoDB ObjectId"
                      },
                      "title": {
                        "type": "string",
                        "example": "Match Recap"
                      },
                      "content": {
                        "type": "string",
                        "example": "A detailed recap of yesterday's game."
                      },
                      "sportSlug": {
                        "type": "string",
                        "example": "soccer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/clubs": {
      "get": {
        "summary": "Fetch clubs with optional filtering, location, and pagination",
        "tags": [
          "Clubs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sport",
            "schema": {
              "type": "string"
            },
            "description": "Filter clubs by sport slug (case-insensitive). Use \"all\" for no filter."
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Filter clubs by name (partial match)."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Pagination page number."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of items per page."
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string",
              "example": "40.7128,-74.0060"
            },
            "description": "Latitude and longitude to filter nearby clubs."
          },
          {
            "in": "query",
            "name": "distance",
            "schema": {
              "type": "number",
              "default": 0
            },
            "description": "Distance in miles to filter clubs near the location."
          }
        ],
        "responses": {
          "200": {
            "description": "Array of clubs merged with overrides."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/events": {
      "post": {
        "summary": "Add or update events",
        "tags": [
          "Clubs"
        ],
        "description": "Accepts a single event object or an array of events.\nEvents are identified by a unique `hash`. Existing events with the same `hash` are deleted before inserting.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": null
          }
        },
        "responses": {
          "200": {
            "description": "Successfully inserted events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "insertedCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (empty array or missing hash)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/articles": {
      "get": {
        "summary": "Get all articles",
        "description": "Retrieve all articles from the database, sorted by name in ascending order.",
        "tags": [
          "Articles"
        ],
        "responses": {
          "200": {
            "description": "List of articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "MongoDB ObjectId"
                      },
                      "title": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or update an article",
        "description": "Insert a new article or update an existing one by slug or name (upsert).",
        "tags": [
          "Articles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "New Article"
                  },
                  "content": {
                    "type": "string",
                    "example": "This is the content of the article."
                  },
                  "slug": {
                    "type": "string",
                    "example": "new-article"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created or updated"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}